+++ solution-pharmacy/models/__init__.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 
 from . import models
-from . import products_quatator
+from . import products_quotator
+from . import products_quotator_lines
 from . import res_partner
