+++ solution-pharmacy/models/res_partner.py
@@ -10,4 +10,3 @@ class RestPartnerType(models.Model):
                                    selection=[('specialist', 'Specialist'), ('distributor', 'Distributor'), ('patient', 'Patient')],
                                    requeried=True,
                                    store=True)
-
