+++ solution-pharmacy/security/ir.model.access.csv
@@ -1,3 +1,4 @@
 id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
 access_solution-pharmacy_solution-pharmacy,solution-pharmacy.solution-pharmacy,model_solution-pharmacy_solution-pharmacy,base.group_user,1,1,1,1
-access_solution_pharmacy_quatator_user,solution-pharmacy.solution.pharmacy.quatator,model_solution_pharmacy_quatator,base.group_user,1,1,1,1
+access_solution_pharmacy_quotator_user,solution-pharmacy.solution.pharmacy.quotator,model_solution_pharmacy_quotator,base.group_user,1,1,1,1
+access_solution_pharmacy_quotator_lines_user,solution-pharmacy.solution.pharmacy.quotator.lines,model_solution_pharmacy_quotator_lines,base.group_user,1,1,1,1
